\docType{package}
\name{subdetect-package}
\alias{subdetect}
\alias{subdetect-package}
\title{Detect Subgroup with an Enhanced Treatment Effect.}
\description{
A test for the existence of a subgroup with enhanced treatment effect and estimation of the subgroup if it exists. In addition, a sample size calculation procedure for this subgroup detection test.
}
\details{
\itemize{
 \item{\code{\link[subdetect]{subgroup_detect}}:}{ test for the existence of a subgroup and estimation of the subgroup.}
 \item{\code{\link[subdetect]{sample_size}}:}{estimate of the required sample size for subgroup detection test.}
}
}
\author{
Ailin Fan and Shannon T. Holloway,

Maintainer: Shannon T. Holloway \email{shannon.t.holloway@gmail.com}
}
\references{
Ailin Fan, Rui Song, and Wenbin Lu, (2016).
Change-plane analysis for subgroup detection and sample size calculation, 
Journal of the American Statistical Association, in press.
}

